quinto estágio -2 - projeto em fase final. Logs direcionando possíveis erros, janelas redimensionadas, falta implementar testes e migrar banco de dados para o MySQL. Preparando para migrar para o MySQL

+++ accounts/views.py
@@ -6,7 +6,7 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.models import User as AuthUser  # Evitando confusão com nosso User model
 from django.core.serializers import serialize
 from django.db.migrations import serializer
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from django.views import View
 from django.views.generic import TemplateView
 from rest_framework import generics, status
@@ -70,7 +70,7 @@ class HomePageView(LoginRequiredMixin, TemplateView):
 class UserListView(generics.ListCreateAPIView):
     queryset = User.objects.all()
     serializer_class = UserSerializer
-    permission_classes = [AllowAny]  # Você pode ajustar as permissões conforme necessário
+    permission_classes = [IsAuthenticated]  # Você pode ajustar as permissões conforme necessário
     template_name = 'users-list.html'
 
     def get(self, request, *args, **kwargs):
@@ -81,31 +81,65 @@ class UserListView(generics.ListCreateAPIView):
 class UserDetailView(generics.RetrieveUpdateDestroyAPIView):
     queryset = User.objects.all()
     serializer_class = UserSerializer
-    permission_classes = [IsAuthenticated]  # Apenas usuários autenticados tem acesso
+    permission_classes = [IsAuthenticated]
 
     def get_object(self):
         obj = super().get_object()
         if self.request.user.is_superuser or obj == self.request.user:
             return obj
-        raise PermissionDenied("Você não tem permissão para acessar este usuário.")
+        # Se o usuário não tiver permissão, exibe a mensagem e redireciona para a lista de usuários
+        messages.error(self.request, 'Você não tem permissão para acessar esse usuário.')
+        return None
 
     def get(self, request, *args, **kwargs):
         user = self.get_object()
+        # Redireciona se o usuário não tiver permissão
+        if not user:
+            return redirect('users-list')
         return render(request, 'user-detail.html', {'user': user})
+
+    def post(self, request, *args, **kwargs):
+        user = self.get_object()
+        if not user:
+            return redirect('users-list')
+
+        # Atualização dos dados do usuário
+        user.first_name = request.POST.get('first_name')
+        user.email = request.POST.get('email')
+        user.save()
+
+        messages.success(request, 'Usuário atualizado com sucesso!')
+        return redirect("user-detail", pk=user.pk)
 
     def put(self, request, *args, **kwargs):
         user = self.get_object()
+        if not user:
+            return redirect('users-list')
+
         serializer = self.get_serializer(user, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request):
+    def delete(self, request, *args, **kwargs):
         user = self.get_object()
-        user.delete()
-        return Response(status.HTTP_204_NO_CONTENT)
+        if not user:
+            return redirect('users-list')
 
+        try:
+            user.delete()
+            messages.success(request, 'Usuário deletado com sucesso!')
+            return redirect('users-list')
+        except Exception as e:
+            messages.error(request, f'Erro ao deletar o usuário: {str(e)}')
+            return redirect('user-detail', pk=user.id)
+
+    def dispatch(self, request, *args, **kwargs):
+        # Verifica o método HTTP
+        if request.method == 'POST' and request.POST.get('_method') == 'DELETE':
+            return self.delete(request, *args, **kwargs)
+        return super().dispatch(request, *args, **kwargs)
 
 
 class LogoutView(View):
